chore: slices permission to services, types and funcs files
